+++ raveandroid/src/main/java/com/flutterwave/raveandroid/di/components/ApplicationComponents.java
@@ -1,7 +1,7 @@
-package com.flutterwave.raveandroid.components;
+package com.flutterwave.raveandroid.di.components;
 
-import com.flutterwave.raveandroid.modules.MpesaModule;
-import com.flutterwave.raveandroid.modules.NetworkModule;
+import com.flutterwave.raveandroid.di.modules.MpesaModule;
+import com.flutterwave.raveandroid.di.modules.NetworkModule;
 import com.flutterwave.raveandroid.mpesa.MpesaFragment;
 
 import javax.inject.Singleton;
