+++ raveandroid/src/main/java/com/flutterwave/raveandroid/mpesa/MpesaFragment.java
@@ -22,11 +22,11 @@ import com.flutterwave.raveandroid.RavePayActivity;
 import com.flutterwave.raveandroid.RavePayInitializer;
 import com.flutterwave.raveandroid.Utils;
 import com.flutterwave.raveandroid.ViewObject;
-import com.flutterwave.raveandroid.components.ApplicationComponents;
-import com.flutterwave.raveandroid.components.DaggerApplicationComponents;
 import com.flutterwave.raveandroid.data.ApiService;
-import com.flutterwave.raveandroid.modules.MpesaModule;
-import com.flutterwave.raveandroid.modules.NetworkModule;
+import com.flutterwave.raveandroid.di.components.ApplicationComponents;
+import com.flutterwave.raveandroid.di.components.DaggerApplicationComponents;
+import com.flutterwave.raveandroid.di.modules.MpesaModule;
+import com.flutterwave.raveandroid.di.modules.NetworkModule;
 
 import java.util.HashMap;
